Add files via upload
Old text sentiment analysis techniques often neglect the
context when it is used in the expression of features. In
this research, we have implemented a scalable and robust
CNN (Convolutional neural network) to address the issue
of text-based film review sentiment analysis. The proposed
method classifies the sentiment of sentences in the Rotten
Tomatoes data set. Basically, it is a multi-class classification,
which means that the data set have more than 2 classes for
classifying the movies. Finally, we are evaluating the model
on testing dataset to check how it performs by calculating
Accuracy, Recall, Precision and figure of merit measures.
Moreover, this paper also discuss about the previously
implemented research work on sentiment analysis which uses
different methods on various dataset.